feat: 클래스 이름 수정 및 메시저 전송 메서드 추가
